feat: expose watcher plugin test harness
